Add files via upload
Fixed bug where particles are skipped during the "for particle in particles" loops:

If a particle object is destroyed during the loop and removed from the particles list, the following objects all have their indexes reduced by one as the list shrinks. This means that the for loop skips the following particle out when it iterates. This was fixed by making sure the loops iterate backwards from the final element in the list instead, using "for particle in reversed(particles))". This means that all changes to element indexes do not affect the particles which have not yet been processed in the loop.

Also added some extra plotting, with a time structure plot, and a few other tweaks here and there