Add files via upload
Hadn't backed up for a while so I've added this one. The code is nearing completion, and the ability to switch efields is currently being implemented, Better plotting and trajectory stuff has been added, and the code now works with fields of different sizes and with different step_sizes (before it crashed whenever you changed either of these).