Add files via upload
This is the heavily (and badly) modified version of the code used to generate plots for IBIC 2019 conference work. The code is broken into seperate cells to allow different plots to be generated once tracking is complete, and in some cases sets of data are entered into the code directly as lists. However its not very usable or scaleable. From this point on the majority of plots and hard-coded data will be removed, but this version will be left here so that the capabilities are not permanently lost if needed.